fixed tests compilation
